requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
openai-whisper==20231117
torch==2.3.0
transformers==4.41.2
datasets==2.19.2
scikit-learn==1.5.0
seaborn==0.13.2
matplotlib==3.9.0
gTTS==2.5.1
pydub==0.25.1
numpy==1.26.4
pandas==2.2.2
jiwer==3.0.4
pyyaml==6.0.1
accelerate==0.30.0
